Use `GetLayoutManager` with `TabStripRegionView`
`TabStripRegionView` used to have a `layout_manager_` data field,
however this has been changed to rely on the layout manager already
present in `View`.

This change uses that accessor through a cast, as is done in other
places in upstream, to keep the same changes to the layout manager as
before.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/0139ceffb2f8e1f64b7c30834e57d5793e529ed7

commit 0139ceffb2f8e1f64b7c30834e57d5793e529ed7
Author: Peter Kasting <pkasting@chromium.org>
Date:   Mon Feb 5 18:24:38 2024 +0000

    Use kViewIgnoredByLayoutKey everywhere.

    This removes LayoutManagerBase::*ChildViewIgnoredByLayout(), which is
    no longer necessary.

    Bug: 1267319