[Android] Remove semi-asynchronous callback from Java API for QR code generation
Chromium change:
https://source.chromium.org/chromium/chromium/src/+/8fd2437a2c41996cd56a09df78dbec8db8cd84a3

Remove semi-asynchronous callback from Java API for QR code generation.

This is a follow-up to https://crrev.com/c/5278872 where we removed
the semi-asynchronous callback from the C++ code and replaced it with
directly returning the generated QR code.  This CL does the same thing
on Java side.

Bug: 40263739