Moving `searchable-drop-down` into settings
This component got moved into Print Preview in upstream, to not be used
outside it. This change adds a copy of it into brave settings, for the
wallet code. This is a temporary change as the `iron-dropdown` has been
deprecated, and some additional rework will be need going forward.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/7246277959ff32fb2768d672e36833070bffc9d0

commit 7246277959ff32fb2768d672e36833070bffc9d0
Author: rbpotter <rbpotter@chromium.org>
Date:   Fri Mar 8 23:59:01 2024 +0000

    WebUI: Move cr-searchable-drop-down to only user

    Move cr-searchable-drop-down to Print Preview. This element is only
    used in CrOS Print Preview, and is not allowed on other platforms due
    to use of deprecated iron-dropdown. Moving it out of the shared
    resources may help to discourage any further use.

    Rename it to searchable_drop_down_cros so that the CrOS Printing team
    can own it.

    Bug: 40943652