First steps towards Omaha 4 on macOS
This commit adds an implementation that replaces Sparkle by Omaha 4 for
on-demand updates. It is hidden behind a default-disabled GN arg and
is highly experimental: The user must manually install Omaha 4 by
executing a command. They must also manually create a JSON file that
tells Omaha that Brave is installed. Once they do that, the
implementation can update to a new version via brave://settings/help.
The associated "Brave was updated - relaunch?" button does not work.