[iOS] Update secure content state update flow
This change moves all of the logic that used to happen in each KVO observation into one unified location so that it has all the data available each update. It also only updates the tab content state once per update which should fix possible races when updating the URL bar in other locations.